+++ pkg/gokube/gokube.go
@@ -89,7 +89,7 @@ func ReadConfig(verbose bool) error {
 }
 
 // WriteConfig ...
-func WriteConfig(gokubeVersion string, kubernetesVersion string) error {
+func WriteConfig(gokubeVersion string, kubernetesVersion string, containerRuntime string) error {
 	configPath := utils.GetUserHome() + string(os.PathSeparator) + ".gokube"
 	configFile := "config"
 	configFilePath := configPath + string(os.PathSeparator) + "config.yaml"
@@ -110,6 +110,7 @@ func WriteConfig(gokubeVersion string, kubernetesVersion string) error {
 	viper.SetConfigType("yaml")
 	viper.Set("gokube-version", gokubeVersion)
 	viper.Set("kubernetes-version", kubernetesVersion)
+	viper.Set("container-runtime", containerRuntime)
 	err := viper.WriteConfig()
 	if err != nil {
 		return err